+++ cardinal-components-base/src/main/java/dev/onyxstudios/cca/internal/base/ComponentRegistrationInitializer.java
@@ -20,27 +20,10 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package dev.onyxstudios.cca.api.v3.item;
-
-import dev.onyxstudios.cca.internal.base.ComponentRegistrationInitializer;
-import net.minecraft.item.ItemStack;
+package dev.onyxstudios.cca.internal.base;
 
 /**
- * Entrypoint getting invoked to register <em>static</em> item component factories.
- *
- * <p>The entrypoint is exposed as either {@code "cardinal-components"} or {@code "cardinal-components-item"} in the mod json and runs for any environment.
- * It usually executes right before the first {@link ItemStack} instance is created.
- *
- * @since 2.4.0
+ * Internal marker for {@code "cardinal-components"} entrypoints
  */
-public interface ItemComponentInitializer extends ComponentRegistrationInitializer {
-    /**
-     * Called to register component factories for statically declared component types.
-     *
-     * <p><strong>The passed registry must not be held onto!</strong> Static component factories
-     * must not be registered outside of this method.
-     *
-     * @param registry an {@link ItemComponentFactoryRegistry} for <em>statically declared</em> components
-     */
-    void registerItemComponentFactories(ItemComponentFactoryRegistry registry);
+public interface ComponentRegistrationInitializer {
 }