+++ cardinal-components-chunk/src/main/java/dev/onyxstudios/cca/internal/chunk/StaticChunkComponentPlugin.java
@@ -29,7 +29,6 @@ import dev.onyxstudios.cca.api.v3.component.ComponentKey;
 import dev.onyxstudios.cca.internal.base.DynamicContainerFactory;
 import dev.onyxstudios.cca.internal.base.asm.StaticComponentPluginBase;
 import nerdhub.cardinal.components.api.component.Component;
-import net.fabricmc.loader.api.FabricLoader;
 import net.fabricmc.loader.api.entrypoint.EntrypointContainer;
 import net.minecraft.world.chunk.Chunk;
 
@@ -47,7 +46,7 @@ public final class StaticChunkComponentPlugin extends StaticComponentPluginBase<
 
     @Override
     protected Collection<EntrypointContainer<ChunkComponentInitializer>> getEntrypoints() {
-        return FabricLoader.getInstance().getEntrypointContainers("cardinal-components-chunk", ChunkComponentInitializer.class);
+        return getComponentEntrypoints("cardinal-components-chunk", ChunkComponentInitializer.class);
     }
 
     @Override